requirements.txt
PyYAML>=6.0
openai>=1.40.0
tiktoken>=0.7.0
streamlit>=1.36.0
fastapi>=0.111.0
uvicorn>=0.30.0
